Change in Watermark classes - Change from float to double to maintain larger values

+++ source/src/main/java/com/linkedin/uif/source/extractor/watermark/DateWatermark.java
@@ -106,7 +106,7 @@ public class DateWatermark implements Watermark {
     }
 
     int dayInterval = hourInterval / 24;
-    int totalHours = (int) Math.ceil(((float) diffInMilliSecs / (60 * 60 * 1000)));
+    int totalHours = (int) Math.ceil(((double) diffInMilliSecs / (60 * 60 * 1000)));
     long totalIntervals = (long) Math.ceil((float) totalHours / (dayInterval * 24));
     if (totalIntervals > maxIntervals) {
       hourInterval = (int) Math.ceil((float) totalHours / maxIntervals);
