+++ source/src/main/java/com/linkedin/uif/source/extractor/watermark/HourWatermark.java
@@ -103,7 +103,7 @@ public class HourWatermark implements Watermark {
       return 0;
     }
 
-    int totalHours = (int) Math.ceil(((float) diffInMilliSecs / (60 * 60 * 1000)));
+    int totalHours = (int) Math.ceil(((double) diffInMilliSecs / (60 * 60 * 1000)));
     long totalIntervals = (long) Math.ceil((float) totalHours / hourInterval);
     if (totalIntervals > maxIntervals) {
       hourInterval = (int) Math.ceil((float) totalHours / maxIntervals);
